refactor: extract insight savings rate chart
